location: Fix CMock unit test
Now that we are usign DT macros for nRF70 Wi-Fi, the CMock which runs on
native posix board fails as there is no Wi-Fi interface on native posix
and no mocking for DT APIs.

Move the Wi-Fi interface get to a separate function so that it can be
mocked and also as the DT APIs are macros (compile time) they cannot be
mocked, so, add a DTS overlay for native posix board to fix the build.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>