boards: nrf7002dk: port to HWMv2
This commit converts the nrf7002dk to HWMv2.
The emulated nrf7002dk_nrf7001 is made as a board variant.

This means that the board target nrf7002dk/nrf5340/cpuapp builds for the
normal nrf7002 companion IC.

To build for the nrf7001 in emulation mode, then /nrf7001 must be added,
like this: nrf7002dk/nrf5340/cpuapp/nrf7001.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>