applications: serial_lte_modem: PPP to work from AT UART
Making PPP to work from the UART used for AT commands.
This means that the SLM UART is first utilized for AT and
once AT+CFUN=1 and AT#XPPP=1 have been done, it's used for PPP.
Going back from PPP to AT with the same UART is not possible,
or requires a reboot.

Jira: SLM-91

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>